# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: OTP WhatsApp Auth API
  description: Contract-first authentication API using WhatsApp OTP
  version: 1.0.0

servers:
  - url: /api/v1

tags:
  - name: Auth
    description: Authentication endpoints
  - name: User
    description: User profile endpoints
  - name: Seat
    description: Theatre seat endpoints
  - name: Admin
    description: Admin reservation management
  - name: SuperAdmin
    description: Super Admin management
  - name: Public
    description: Public endpoints (no authentication required)

paths:
  /auth/request-otp:
    post:
      tags:
        - Auth
      operationId: requestOtp
      summary: Register/login - sends OTP to user's WhatsApp
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpRequest'
      responses:
        '200':
          description: OTP sent successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /auth/verify-otp:
    post:
      tags:
        - Auth
      operationId: verifyOtp
      summary: Verify OTP and receive JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OtpVerifyRequest'
      responses:
        '200':
          description: OTP verified, JWT issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid or expired OTP
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: OTP mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /user/me:
    get:
      tags:
        - User
      operationId: getMe
      summary: Get the authenticated user's profile
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /user/name:
    put:
      tags:
        - User
      operationId: updateName
      summary: Set or update the user's name (required on first login)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateNameRequest'
      responses:
        '200':
          description: Name updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /user/reservations:
    get:
      tags:
        - User
      operationId: getUserReservations
      summary: Get all reservations for the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of user reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDto'
        '401':
          description: Unauthorized

  /seats/reserve:
    post:
      tags:
        - Seat
      operationId: reserveSeats
      summary: Reserve a list of theatre seats with status PENDING for 2 hours
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReserveSeatsRequest'
      responses:
        '200':
          description: Seats reserved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '400':
          description: One or more seats are already reserved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized

  /seats/reserved:
    get:
      tags:
        - Seat
      operationId: getReservedSeats
      summary: Get all reserved seats (PENDING or CONFIRMED)
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of reserved seats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SeatDto'
        '401':
          description: Unauthorized

  /admin/reservations:
    get:
      tags:
        - Admin
      operationId: getAllReservations
      summary: Admin - retrieve all reservations with seats
      security:
        - bearerAuth: []
      responses:
        '200':
          description: All reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReservationDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /admin/reservations/{id}/confirm:
    post:
      tags:
        - Admin
      operationId: confirmReservation
      summary: Admin - confirm a pending reservation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Reservation confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '400':
          description: Reservation not found or not confirmable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /admin/balance:
    get:
      tags:
        - Admin
      operationId: getAdminBalance
      summary: Admin - get current balance earned from confirmed reservations
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Admin balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /admin/transfers:
    post:
      tags:
        - Admin
      operationId: createTransfer
      summary: Admin - create a transfer to super admin
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransferRequest'
      responses:
        '200':
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferDto'
        '400':
          description: Invalid amount or insufficient balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /super-admin/transfers:
    get:
      tags:
        - SuperAdmin
      operationId: getAllTransfers
      summary: Super Admin - get all transfers from admins
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of all transfers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransferDto'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

  /public/ticket/{token}:
    get:
      tags:
        - Public
      operationId: getTicketByToken
      summary: Public - get reservation details by ticket token (for theatre entry verification)
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Reservation details for the valid confirmed ticket
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationDto'
        '404':
          description: Ticket not found or reservation not yet confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /super-admin/transfers/{id}/confirm:
    post:
      tags:
        - SuperAdmin
      operationId: confirmTransfer
      summary: Super Admin - confirm a pending transfer
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Transfer confirmed, money moved from admin to super admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferDto'
        '400':
          description: Transfer not found or not confirmable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
        '403':
          description: Forbidden

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OtpRequest:
      type: object
      required:
        - phone
      properties:
        phone:
          type: string
          example: "+201234567890"
          description: Phone number in international format

    OtpVerifyRequest:
      type: object
      required:
        - phone
        - otp
      properties:
        phone:
          type: string
          example: "+201234567890"
          description: Phone number used during OTP request
        otp:
          type: string
          example: "123456"
          description: 6-digit OTP received on WhatsApp

    AuthResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT access token
        firstLogin:
          type: boolean
          description: True if this is the user's first login (name not yet set)
        user:
          $ref: '#/components/schemas/UserDto'

    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        phone:
          type: string
        firstLogin:
          type: boolean
        role:
          type: string
          enum: [USER, ADMIN, SUPER_ADMIN]
          description: Role assigned to the user
        balance:
          type: number
          description: Current balance (applicable to admin and super admin users)

    SeatDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        label:
          type: string
        status:
          type: string
          enum: [AVAILABLE, PENDING, CONFIRMED]

    ReservationDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          type: string
          enum: [PENDING, CONFIRMED]
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        totalAmount:
          type: number
          description: Total amount charged (ticketPrice × seats count), set on confirmation
        notes:
          type: string
          description: Notes in Arabic for the reservation
        paymentLink:
          type: string
          description: WhatsApp link to contact the assigned admin with payment details
        confirmedBy:
          type: string
          description: Name of the admin who confirmed the reservation, empty if not yet confirmed
        assignedTo:
          type: string
          description: Name of the admin assigned to this reservation at creation time
        ticketToken:
          type: string
          description: Unique HMAC-signed token for theatre entry verification (generated on confirmation)
        seats:
          type: array
          items:
            $ref: '#/components/schemas/SeatDto'
        user:
          $ref: '#/components/schemas/UserDto'

    ReserveSeatsRequest:
      type: object
      required:
        - seatLabels
      properties:
        seatLabels:
          type: array
          items:
            type: string
          example: ["STAGE-A11", "STAGE-A13"]
        notes:
          type: string
          description: Optional notes in Arabic for the reservation

    UpdateNameRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: John Doe
          description: The user's full name

    MessageResponse:
      type: object
      properties:
        message:
          type: string

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string

    TransferDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        admin:
          $ref: '#/components/schemas/UserDto'
        amount:
          type: number
        status:
          type: string
          enum: [PENDING, CONFIRMED]
        createdAt:
          type: string
          format: date-time

    CreateTransferRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          description: Amount to transfer to super admin

    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
          description: Current balance